Standardize Add/Edit subscription forms
