+++ dashboard/src/pages/admin/subscriptions/AdminSubscriptionsPage.tsx
@@ -1,8 +1,8 @@
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Loader from "components/Loader";
-import SubscriptionAdd from "components/SubscriptionAdd";
 import SubscriptionList from "components/SubscriptionList";
+import SubscriptionAdd from "components/subscriptions/SubscriptionAdd";
 import { useAppDispatch, useAppSelector } from "helpers/hooks";
 import { useEffect } from "react";
 import { Button, Card, Col, Dropdown, Form, Row, Stack } from "react-bootstrap";
